Make MainArgs platform-specific
