+++ crates/cef-ui/src/cef/mod.rs
@@ -17,7 +17,6 @@ mod ime;
 mod keyboard_handler;
 mod life_span_handler;
 mod macros;
-mod main_args;
 mod navigation_entry;
 mod platform;
 mod process;
@@ -58,7 +57,6 @@ pub use ime::*;
 pub use keyboard_handler::*;
 pub use life_span_handler::*;
 pub use macros::*;
-pub use main_args::*;
 pub use navigation_entry::*;
 pub use platform::*;
 pub use process::*;
