+++ crates/cef-ui/src/cef/platform/macos.rs
@@ -1,9 +1,57 @@
 use crate::{
-    bindings::{cef_event_handle_t, cef_string_t, cef_window_handle_t, cef_window_info_t},
+    bindings::{
+        cef_event_handle_t, cef_main_args_t, cef_string_t, cef_window_handle_t, cef_window_info_t
+    },
     free_cef_string, CefString, Rect
 };
 use anyhow::{anyhow, Error, Result};
-use std::{ffi::c_int, mem::zeroed};
+use std::{
+    env::args,
+    ffi::{c_char, c_int, CString},
+    mem::zeroed
+};
+
+/// Structure representing CefExecuteProcess arguments.
+#[derive(Debug)]
+#[allow(dead_code)]
+pub struct MainArgs {
+    // We must keep the CString vector alive
+    // for the pointer vector to remain valid.
+    args: Vec<CString>,
+    argv: Vec<*const c_char>,
+    cef:  cef_main_args_t
+}
+
+impl MainArgs {
+    /// Try and create a new MainArgs from an iterator of strings.
+    pub fn new() -> Result<Self> {
+        let args = args()
+            .into_iter()
+            .map(|arg| CString::new(arg))
+            .collect::<Result<Vec<CString>, _>>()?;
+        let args = args
+            .into_iter()
+            .map(|arg| CString::new(arg))
+            .collect::<Result<Vec<CString>, _>>()?;
+
+        let argv = args
+            .iter()
+            .map(|arg| arg.as_ptr())
+            .collect::<Vec<*const c_char>>();
+
+        let cef = cef_main_args_t {
+            argc: argv.len() as i32,
+            argv: argv.as_ptr() as *mut *mut c_char
+        };
+
+        Ok(Self { args, argv, cef })
+    }
+
+    /// Converts to the raw cef type.
+    pub fn as_raw(&self) -> &cef_main_args_t {
+        &self.cef
+    }
+}
 
 /// Native window handle.
 #[derive(Clone)]
